add xml detail group fragment and change name edit group fragmet
