+++ app/src/main/res/drawable/ic_money.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="29dp"
+    android:height="27dp"
+    android:viewportWidth="29"
+    android:viewportHeight="27">
+  <path
+      android:pathData="M26.625,22.25V23.7083C26.625,25.3125 25.3125,26.625 23.7083,26.625H3.2917C1.6729,26.625 0.375,25.3125 0.375,23.7083V3.2917C0.375,1.6875 1.6729,0.375 3.2917,0.375H23.7083C25.3125,0.375 26.625,1.6875 26.625,3.2917V4.75H13.5C11.8813,4.75 10.5833,6.0625 10.5833,7.6667V19.3333C10.5833,20.9375 11.8813,22.25 13.5,22.25H26.625ZM13.5,19.3333H28.0833V7.6667H13.5V19.3333ZM19.3333,15.6875C18.1229,15.6875 17.1458,14.7104 17.1458,13.5C17.1458,12.2896 18.1229,11.3125 19.3333,11.3125C20.5438,11.3125 21.5208,12.2896 21.5208,13.5C21.5208,14.7104 20.5438,15.6875 19.3333,15.6875Z"
+      android:fillColor="#ffffff"/>
+</vector>
