+++ app/src/main/res/drawable/ic_shop.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="33dp"
+    android:height="29dp"
+    android:viewportWidth="33"
+    android:viewportHeight="29">
+  <path
+      android:pathData="M24.098,11.125L17.7105,1.5583C17.4334,1.15 16.9667,0.9458 16.5,0.9458C16.0334,0.9458 15.5667,1.15 15.2896,1.5729L8.9021,11.125H1.9167C1.1146,11.125 0.4584,11.7812 0.4584,12.5833C0.4584,12.7146 0.473,12.8458 0.5167,12.9771L4.2209,26.4958C4.5563,27.7208 5.6792,28.625 7.0209,28.625H25.9792C27.3209,28.625 28.4438,27.7208 28.7938,26.4958L32.498,12.9771L32.5417,12.5833C32.5417,11.7812 31.8855,11.125 31.0834,11.125H24.098ZM12.125,11.125L16.5,4.7083L20.875,11.125H12.125ZM16.5,22.7916C14.8959,22.7916 13.5834,21.4791 13.5834,19.875C13.5834,18.2708 14.8959,16.9583 16.5,16.9583C18.1042,16.9583 19.4167,18.2708 19.4167,19.875C19.4167,21.4791 18.1042,22.7916 16.5,22.7916Z"
+      android:fillColor="#ffffff"/>
+</vector>
