+++ app/src/main/res/layout/activity_detail_group.xml
@@ -16,14 +16,14 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentDetailGroup"
+        android:name="com.homies.homies.DetailGroupFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_marginTop="65dp"
-        android:name="com.homies.homies.DetailGroupFragment"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/toolbarDetail" />
+        app:layout_constraintTop_toBottomOf="@+id/toolbarDetail"/>
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
