+++ app/src/main/res/layout/fragment_detail_group.xml
@@ -1,85 +1,47 @@
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".DetailGroup">
+    android:layout_height="match_parent">
 
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/ip_groupName"
-        style="@style/Widget.MaterialComponents.TextInputLayout.FilledBox"
-        android:layout_width="match_parent"
+    <Button
+        android:id="@+id/Btn_expenses"
+        android:layout_width="300dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="32dp"
-        android:layout_marginEnd="20dp"
+        android:layout_centerInParent="true"
+        android:layout_marginTop="100dp"
+        android:background="@drawable/btn_custom"
+        android:fontFamily="@font/montserrat_regular"
         android:padding="1dp"
-        app:boxStrokeWidth="0dp"
-        app:boxStrokeWidthFocused="0dp"
-        app:errorTextAppearance="@style/error_appearance"
+        android:text="@string/btn_expenses"
+        android:textAllCaps="false"
+        android:textColor="@color/white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:startIconDrawable="@drawable/ic_group_name_home"
-        app:startIconTint="@color/primaryColor">
+        android:drawableLeft="@drawable/ic_money"/>
 
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/textInputGroupName"
-            android:layout_width="match_parent"
-            android:layout_height="61dp"
-            android:layout_marginStart="0dp"
-            android:layout_marginTop="0dp"
-            android:layout_marginEnd="0dp"
-            android:layout_marginBottom="0dp"
-            android:background="@drawable/rounded_edittext"
-            android:fontFamily="@font/montserrat_regular"
-            android:inputType="text"
-            android:maxLength="50"
-            android:padding="15dp"
-            android:text="@string/groupName"
-            android:textColor="@color/fieldTextColor"
-            android:textSize="15sp" />
-    </com.google.android.material.textfield.TextInputLayout>
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/ip_groupDetail"
-        style="@style/Widget.MaterialComponents.TextInputLayout.FilledBox"
-        android:layout_width="match_parent"
-        android:layout_height="100dp"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="10dp"
-        android:layout_marginEnd="20dp"
+    <Button
+        android:id="@+id/Btn_shoppingList"
+        android:layout_width="300dp"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:layout_marginTop="20dp"
+        android:background="@drawable/btn_custom"
+        android:fontFamily="@font/montserrat_regular"
         android:padding="1dp"
-        android:textAlignment="textStart"
-        app:boxStrokeWidth="0dp"
-        app:boxStrokeWidthFocused="0dp"
-        app:errorTextAppearance="@style/error_appearance"
+        android:text="@string/btn_shopping_list"
+        android:textAllCaps="false"
+        android:textColor="@color/white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/ip_groupName"
-        app:startIconDrawable="@drawable/ic_mode_edit"
-        app:startIconTint="@color/primaryColor">
-
-        <com.google.android.material.textfield.TextInputEditText
-            android:id="@+id/textInputDetailGroupName"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginStart="0dp"
-            android:layout_marginTop="0dp"
-            android:layout_marginEnd="0dp"
-            android:layout_marginBottom="0dp"
-            android:background="@drawable/rounded_edittext"
-            android:fontFamily="@font/montserrat_regular"
-            android:inputType="text"
-            android:maxLength="50"
-            android:padding="15dp"
-            android:text="@string/groupDescription"
-            android:textColor="@color/fieldTextColor"
-            android:textSize="15sp" />
-    </com.google.android.material.textfield.TextInputLayout>
+        app:layout_constraintTop_toBottomOf="@+id/Btn_expenses"
+        android:drawableLeft="@drawable/ic_shop"/>
 
     <ListView
-        android:id="@+id/userList"
+        android:id="@+id/taskList"
         android:name="com.homies.homies.DetailGroupFragment"
         android:layout_width="match_parent"
         android:layout_height="150dp"
@@ -88,58 +50,39 @@
         android:layout_marginEnd="20dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/ip_groupDetail" />
+        app:layout_constraintTop_toBottomOf="@+id/Btn_shoppingList"/>
+
 
     <Button
-        android:id="@+id/btnAddUser"
-        android:layout_width="150dp"
+        android:id="@+id/btn_addTask"
+        android:layout_width="100dp"
         android:layout_height="wrap_content"
         android:layout_centerInParent="true"
         android:layout_marginTop="20dp"
         android:background="@drawable/btn_custom"
         android:fontFamily="@font/montserrat_regular"
         android:padding="1dp"
-        android:text="@string/añadir_usuario"
-        android:textAllCaps="false"
-        android:textColor="@color/primaryTextColor"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/userList" />
-
-    <Button
-        android:id="@+id/btnLeaveGroup"
-        android:layout_width="300dp"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:layout_marginTop="25dp"
-        android:background="@drawable/btn_error"
-        android:fontFamily="@font/montserrat_regular"
-        android:padding="1dp"
-        android:text="@string/Btn_leave_group"
+        android:text="@string/btn_add_task"
         android:textAllCaps="false"
         android:textColor="@color/white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btnAddUser" />
-
+        app:layout_constraintTop_toBottomOf="@+id/taskList" />
 
     <Button
-        android:id="@+id/btnDeleteGroup"
+        android:id="@+id/btn_editGroup"
         android:layout_width="300dp"
         android:layout_height="wrap_content"
         android:layout_centerInParent="true"
-        android:layout_marginTop="10dp"
-        android:background="@drawable/btn_error"
+        android:layout_marginTop="20dp"
+        android:background="@drawable/btn_custom"
         android:fontFamily="@font/montserrat_regular"
         android:padding="1dp"
-        android:text="@string/delete_group"
+        android:text="@string/btn_editGroup"
         android:textAllCaps="false"
         android:textColor="@color/white"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btnLeaveGroup" />
-
-
-
+        app:layout_constraintTop_toBottomOf="@+id/btn_addTask" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
